:recycle: Refactorin code
