+++ src/main/java/com/moviestore/controller/dto/MovieDto.java
@@ -4,23 +4,23 @@ import java.util.List;
 import java.util.stream.Collectors;
 
 import com.moviestore.model.Movie;
-import com.moviestore.model.UserMovie;
+import com.moviestore.model.Rent;
 
 public class MovieDto {
 
 	private Long id;
 	private String name;
-	
+
 	public MovieDto(Movie movie) {
 		this.id = movie.getId();
 		this.name = movie.getTitle();
 	}
-	
-	public MovieDto(UserMovie usermovie) {
-		this.id = usermovie.getMovie().getId();
-		this.name = usermovie.getMovie().getTitle();
+
+	public MovieDto(Rent rent) {
+		this.id = rent.getMovies().getId();
+		this.name = rent.getMovies().getTitle();
 	}
-	
+
 	public Long getId() {
 		return id;
 	}
@@ -28,13 +28,10 @@ public class MovieDto {
 	public String getName() {
 		return name;
 	}
-	
+
 	public static List<MovieDto> convertMoviesToDto(List<Movie> listMovies) {
 		return listMovies.stream().map(MovieDto::new).collect(Collectors.toList());
 	}
-	
-	public static List<MovieDto> convertUserMoviesDto(List<UserMovie> listUserMovies) {
-		return listUserMovies.stream().map(MovieDto::new).collect(Collectors.toList());
-	}
+
 
 }
