+++ src/main/java/com/moviestore/model/User.java
@@ -1,10 +1,14 @@
 package com.moviestore.model;
 
+import java.util.List;
+
+import javax.persistence.CascadeType;
 import javax.persistence.Column;
 import javax.persistence.Entity;
 import javax.persistence.GeneratedValue;
 import javax.persistence.GenerationType;
 import javax.persistence.Id;
+import javax.persistence.OneToMany;
 import javax.persistence.Table;
 import javax.validation.constraints.NotBlank;
 
@@ -29,7 +33,9 @@ public class User {//implements GrantedAuthority {
 	@NotBlank(message = "Required password field.")
 	@Column(name = "password", nullable = false)
 	private String password;
-		
+	
+	@OneToMany(mappedBy = "user", cascade = CascadeType.ALL, orphanRemoval = true)
+	private List<Rent> rented;
 		
 	public Long getId() {
 		return id;
